--- manager/allocator/cnmallocator/drivers_network_windows.go
@@ -1,17 +1,0 @@
-package cnmallocator
-
-import (
-	"github.com/docker/libnetwork/drivers/overlay/ovmanager"
-	"github.com/docker/libnetwork/drivers/remote"
-	"github.com/docker/swarmkit/manager/allocator/networkallocator"
-)
-
-var initializers = []initializer{
-	{remote.Init, "remote"},
-	{ovmanager.Init, "overlay"},
-}
-
-// PredefinedNetworks returns the list of predefined network structures
-func PredefinedNetworks() []networkallocator.PredefinedNetworkData {
-	return nil
-}
