+++ manager/allocator/helpers/doc.go
@@ -1,0 +1,9 @@
+package helpers
+
+// The helpers package is a package from the allocator that includes common
+// functions used throughout swarmkit and the allocator itself for verification
+// and validation of objects. All of the functions in helpers are stateless and
+// idempotent.
+//
+// Users not using the allocator should import `helpers` and not the allocator
+// packages directly.
