+++ manager/allocator/network/drivers_common.go
@@ -2,6 +2,7 @@ package network
 
 import (
 	"github.com/docker/libnetwork/drvregistry"
+	"github.com/docker/swarmkit/manager/allocator/helpers"
 )
 
 // types common to all driver_platform files
@@ -11,4 +12,17 @@ import (
 type initializer struct {
 	fn    drvregistry.InitFunc
 	ntype string
+}
+
+var (
+	// PredefinedLabel identifies internally allocated swarm networks
+	// corresponding to the node-local predefined networks on the host.
+	PredefinedLabel = helpers.PredefinedLabel
+)
+
+// PredefinedNetworkData contains the minimum set of data needed
+// to create the correspondent predefined network object in the store.
+type PredefinedNetworkData struct {
+	Name   string
+	Driver string
 }
