+++ manager/allocator/network/drivers_network_linux.go
@@ -7,8 +7,10 @@ import (
 	"github.com/docker/libnetwork/drivers/macvlan/mvmanager"
 	"github.com/docker/libnetwork/drivers/overlay/ovmanager"
 	"github.com/docker/libnetwork/drivers/remote"
-	"github.com/docker/swarmkit/manager/allocator/networkallocator"
 )
+
+// NOTE(dperny): these driver names are also indepedently in the helpers
+// package. If you add or remote drivers, you need to do so there as well.
 
 var initializers = []initializer{
 	{remote.Init, "remote"},
@@ -20,7 +22,7 @@ var initializers = []initializer{
 }
 
 // PredefinedNetworks returns the list of predefined network structures
-func PredefinedNetworks() []networkallocator.PredefinedNetworkData {
+func PredefinedNetworks() []PredefinedNetworkData {
 	return []networkallocator.PredefinedNetworkData{
 		{Name: "bridge", Driver: "bridge"},
 		{Name: "host", Driver: "host"},
