+++ manager/allocator/network/drivers_network_windows.go
@@ -3,8 +3,10 @@ package network
 import (
 	"github.com/docker/libnetwork/drivers/overlay/ovmanager"
 	"github.com/docker/libnetwork/drivers/remote"
-	"github.com/docker/swarmkit/manager/allocator/networkallocator"
 )
+
+// NOTE(dperny): these driver names are also indepedently in the helpers
+// package. If you add or remote drivers, you need to do so there as well.
 
 var initializers = []initializer{
 	{remote.Init, "remote"},
@@ -12,6 +14,6 @@ var initializers = []initializer{
 }
 
 // PredefinedNetworks returns the list of predefined network structures
-func PredefinedNetworks() []networkallocator.PredefinedNetworkData {
+func PredefinedNetworks() []PredefinedNetworkData {
 	return nil
 }
